+++ services/src/main/java/ru/dpopkov/knownics/services/QuestionService.java
@@ -1,0 +1,6 @@
+package ru.dpopkov.knownics.services;
+
+import ru.dpopkov.knownics.domain.question.Question;
+
+public interface QuestionService extends BaseService<Question> {
+}